Revert "Doc update"
