+++ .github/workflows/e2e_dapo.yml
@@ -42,7 +42,6 @@ on:
       - main
       - v0.*
     paths:
-      - "**/*.py"
       # Other entrypoints
       - "!examples/*trainer*"
       - "!tests/**"
