+++ .github/workflows/e2e_eval_aime24.yml
@@ -42,7 +42,6 @@ on:
       - main
       - v0.*
     paths:
-      - "**/*.py"
       # Other entrypoints
       - "!*.md"
       - "!docker/**"
