+++ docs/faq/faq.rst
@@ -112,7 +112,7 @@ If in doubt, print this env var in each rank to make sure it is properly set.
 Checkpoints
 ------------------------
 
-If you want to convert the model checkpoint into huggingface safetensor format, please refer to ``verl/model_merger``.
+If you want to convert the model checkpoint into huggingface safetensor format, please refer to ``scripts/model_merger.py``.
 
 
 Triton ``compile_module_from_src`` error
