+++ examples/split_placement/split_monkey_patch.py
@@ -25,11 +25,11 @@ import torch
 from verl import DataProto
 from verl.trainer.ppo.ray_trainer import (
     AdvantageEstimator,
+    _timer,
     apply_kl_penalty,
     compute_advantage,
     compute_data_metrics,
     compute_timing_metrics,
-    marked_timer,
 )
 from verl.utils.metric import reduce_metrics
 
@@ -81,15 +81,15 @@ def fit(self):
             gen_batch = batch.pop(batch_keys=["input_ids", "attention_mask", "position_ids"])
             is_last_step = self.global_steps >= self.total_training_steps
 
-            with marked_timer("step", timing_raw):
+            with _timer("step", timing_raw):
                 # generate a batch
-                with marked_timer("gen", timing_raw):
+                with _timer("gen", timing_raw):
                     gen_batch_output = self.actor_rollout_wg.generate_sequences(gen_batch)
                     timing_raw.update(gen_batch_output.meta_info["timing"])
                     gen_batch_output.meta_info.pop("timing", None)
 
                 if self.config.algorithm.adv_estimator == AdvantageEstimator.REMAX:
-                    with marked_timer("gen_max", timing_raw):
+                    with _timer("gen_max", timing_raw):
                         gen_baseline_batch = deepcopy(gen_batch)
                         gen_baseline_batch.meta_info["do_sample"] = False
                         gen_baseline_output = self.actor_rollout_wg.generate_sequences(gen_baseline_batch)
@@ -120,23 +120,23 @@ def fit(self):
                 batch.meta_info["global_token_num"] = torch.sum(batch.batch["attention_mask"], dim=-1).tolist()
 
                 # recompute old_log_probs
-                with marked_timer("old_log_prob", timing_raw):
+                with _timer("old_log_prob", timing_raw):
                     old_log_prob = self.actor_rollout_wg.compute_log_prob(batch)
                     batch = batch.union(old_log_prob)
 
                 if self.use_reference_policy:
                     # compute reference log_prob
-                    with marked_timer("ref", timing_raw):
+                    with _timer("ref", timing_raw):
                         ref_log_prob = self.ref_policy_wg.compute_ref_log_prob(batch)
                         batch = batch.union(ref_log_prob)
 
                 # compute values
                 if self.use_critic:
-                    with marked_timer("values", timing_raw):
+                    with _timer("values", timing_raw):
                         values = self.critic_wg.compute_values(batch)
                         batch = batch.union(values)
 
-                with marked_timer("adv", timing_raw):
+                with _timer("adv", timing_raw):
                     # compute scores. Support both model and function-based.
                     # We first compute the scores using reward model. Then, we call reward_fn to combine
                     # the results from reward model and rule-based results.
@@ -169,17 +169,17 @@ def fit(self):
 
                 # update critic
                 if self.use_critic:
-                    with marked_timer("update_critic_call", timing_raw):
+                    with _timer("update_critic_call", timing_raw):
                         critic_output = self.critic_wg.update_critic(batch)
 
                 # implement critic warmup
                 if self.config.trainer.critic_warmup <= self.global_steps:
                     # update actor
-                    with marked_timer("update_actor_call", timing_raw):
+                    with _timer("update_actor_call", timing_raw):
                         actor_output = self.actor_rollout_wg.update_actor(batch)
 
                 # NOTE: make sure you set blocking=False in update_actor and update_crtic in the worker class
-                with marked_timer("update_actor_critic", timing_raw):
+                with _timer("update_actor_critic", timing_raw):
                     critic_output = critic_output.get()
                     critic_output_metrics = reduce_metrics(critic_output.meta_info["metrics"])
                     metrics.update(critic_output_metrics)
@@ -190,14 +190,14 @@ def fit(self):
 
                 # validate
                 if self.val_reward_fn is not None and self.config.trainer.test_freq > 0 and (is_last_step or self.global_steps % self.config.trainer.test_freq == 0):
-                    with marked_timer("testing", timing_raw):
+                    with _timer("testing", timing_raw):
                         val_metrics: dict = self._validate()
                         if is_last_step:
                             last_val_metrics = val_metrics
                     metrics.update(val_metrics)
 
                 if self.config.trainer.save_freq > 0 and (is_last_step or self.global_steps % self.config.trainer.save_freq == 0):
-                    with marked_timer("save_checkpoint", timing_raw):
+                    with _timer("save_checkpoint", timing_raw):
                         self._save_checkpoint()
 
             # collect metrics
