+++ recipe/prime/prime_ray_trainer.py
@@ -162,8 +162,8 @@ class RayPRIMETrainer(RayPPOTrainer):
             role_worker_mapping,
             resource_pool_manager,
             ray_worker_group_cls,
-            reward_fn=reward_fn,
-            val_reward_fn=val_reward_fn,
+            reward_fn,
+            val_reward_fn,
             device_name=device_name,
         )
