+++ tests/special_sanity/check_device_api_usage.py
@@ -86,6 +86,6 @@ if __name__ == "__main__":
                     find_invalid_device_management = True
                     break
 
-            print(f"[CHECK] File {path_in_str} is detected for device api usage check, check result: {'success' if not find_invalid_device_management else f'failed, because detect {sk}'}.")
+            print(f"[CHECK] File {path_in_str} is detected for device api usage check, check result: {'success' if not find_invalid_device_management else 'failed'}.")
 
             assert not find_invalid_device_management, f'file {path_in_str} contains .cuda/"cuda"/"nccl" usage, please use api in verl/utils/device.py directly.'
