+++ verl/models/transformers/monkey_patch.py
@@ -211,7 +211,7 @@ def apply_monkey_patch(
     )
 
     if is_trl_available():
-        from trl import AutoModelForCausalLMWithValueHead  # type: ignore
+        from trl import AutoModelForCausalLMWithValueHead
 
         def state_dict(self, *args, **kwargs):
             return torch.nn.Module.state_dict(self, *args, **kwargs)
@@ -221,15 +221,14 @@ def apply_monkey_patch(
 
     # TODO: VLM models only, unify monkey patch to LLM models.
     if model.config.model_type == "qwen2_5_vl":
-        if is_transformers_version_in_range(min_version="4.53.0"):
-            from transformers.models.qwen2_5_vl.modeling_qwen2_5_vl import Qwen2_5_VLAttention
-        else:
-            from transformers.models.qwen2_5_vl.modeling_qwen2_5_vl import Qwen2_5_VLFlashAttention2 as Qwen2_5_VLAttention
+        from transformers.models.qwen2_5_vl.modeling_qwen2_5_vl import (
+            Qwen2_5_VLFlashAttention2,
+        )
 
         if use_remove_padding or ulysses_sp_size > 1:
             from verl.models.transformers.qwen2_vl import ulysses_flash_attn_forward
 
-            Qwen2_5_VLAttention.forward = ulysses_flash_attn_forward
+            Qwen2_5_VLFlashAttention2.forward = ulysses_flash_attn_forward
             print("Monkey patch FlashAttention2.forward in Qwen2.5VL")
 
         if ulysses_sp_size > 1:
@@ -243,15 +242,14 @@ def apply_monkey_patch(
                 patch_vlm_for_ulysses_input_slicing(Qwen2_5_VLModel)
 
     elif model.config.model_type == "qwen2_vl":
-        if is_transformers_version_in_range(min_version="4.53.0"):
-            from transformers.models.qwen2_vl.modeling_qwen2_vl import Qwen2VLAttention
-        else:
-            from transformers.models.qwen2_vl.modeling_qwen2_vl import Qwen2VLFlashAttention2 as Qwen2VLAttention
+        from transformers.models.qwen2_vl.modeling_qwen2_vl import (
+            Qwen2VLFlashAttention2,
+        )
 
         if use_remove_padding or ulysses_sp_size > 1:
             from verl.models.transformers.qwen2_vl import ulysses_flash_attn_forward
 
-            Qwen2VLAttention.forward = ulysses_flash_attn_forward
+            Qwen2VLFlashAttention2.forward = ulysses_flash_attn_forward
             print("Monkey patch FlashAttention2.forward in Qwen2VL")
 
         if ulysses_sp_size > 1:
