+++ verl/trainer/config/ppo_megatron_trainer.yaml
@@ -110,7 +110,6 @@ actor_rollout_ref:
       save_path: null # the path to save the profile result
     load_weight: True
     checkpoint:
-      async_save: False # save checkpoint asynchronously
       # What to include in saved checkpoints
       # with 'hf_model' you can save whole model as hf format, now only use sharded model checkpoint to save space
       save_contents: ['model', 'optimizer', 'extra']
@@ -202,6 +201,7 @@ actor_rollout_ref:
 
     # Multi-turn interaction config for tools or chat.
     multi_turn:
+
       # set to True for multi-turn tool interaction tasks; should set rollout.name to sglang as well
       enable: False
 
@@ -235,9 +235,6 @@ actor_rollout_ref:
       # - ignore_strippable: ignore strippable tokens when checking tokenization sanity
       tokenization_sanity_check_mode: strict
   
-      # Format of the multi-turn interaction. Options: hermes, llama3_json, ...
-      format: hermes
-
     # support logging rollout prob for debugging purpose
     calculate_log_probs: False
     # Nsight system profiler configs
@@ -312,7 +309,6 @@ critic:
     kl_coef: 0.001
   loss_agg_mode: ${actor_rollout_ref.actor.loss_agg_mode}
   checkpoint:
-    async_save: False # save checkpoint asynchronously
     # What to include in saved checkpoints
     # with 'hf_model' you can save whole model as hf format, now only use sharded model checkpoint to save space
     save_contents: ['model', 'optimizer', 'extra']
@@ -396,8 +392,6 @@ trainer:
   nnodes: 1
   n_gpus_per_node: 8
   save_freq: -1
-  save_after_epochs: -1
-  save_best_val_metric: None
   # auto: find the last ckpt to resume. If can't find, start from scratch
   resume_mode: auto # or disable or resume_path if resume_from_path is set
   resume_from_path: null
