+++ verl/utils/checkpoint/checkpoint_manager.py
@@ -15,11 +15,13 @@
 import os
 import random
 import shutil
+import tempfile
 from typing import Union
 
 import numpy as np
 import torch
 import torch.distributed
+from filelock import FileLock
 from omegaconf import DictConfig
 from transformers import PreTrainedTokenizer, ProcessorMixin
 
@@ -47,11 +49,10 @@ class BaseCheckpointManager:
         optimizer: torch.optim.Optimizer,
         lr_scheduler: torch.optim.lr_scheduler.LRScheduler = None,
         processing_class: Union[PreTrainedTokenizer, ProcessorMixin] = None,
-        checkpoint_config: DictConfig = None,
+        checkpoint_contents: DictConfig = None,
     ):
-        self.checkpoint_config = checkpoint_config
-        checkpoint_load_contents = checkpoint_config.get("load_contents", None) if checkpoint_config else None
-        checkpoint_save_contents = checkpoint_config.get("save_contents", None) if checkpoint_config else None
+        checkpoint_load_contents = checkpoint_contents.get("load_contents", None) if checkpoint_contents else None
+        checkpoint_save_contents = checkpoint_contents.get("save_contents", None) if checkpoint_contents else None
         if checkpoint_load_contents is None:
             checkpoint_load_contents = ["model", "optimizer", "extra"]
         if checkpoint_save_contents is None:
@@ -138,6 +139,27 @@ class BaseCheckpointManager:
             if not os.path.exists(abs_path):
                 continue
             shutil.rmtree(abs_path, ignore_errors=True)
+
+    @staticmethod
+    def local_mkdir(path):
+        if not os.path.isabs(path):
+            working_dir = os.getcwd()
+            path = os.path.join(working_dir, path)
+
+        # Using hash value of path as lock file name to avoid long file name
+        lock_filename = f"ckpt_{hash(path) & 0xFFFFFFFF:08x}.lock"
+        lock_path = os.path.join(tempfile.gettempdir(), lock_filename)
+
+        try:
+            with FileLock(lock_path, timeout=60):  # Add timeout
+                # make a new dir
+                os.makedirs(path, exist_ok=True)
+        except Exception as e:
+            print(f"Warning: Failed to acquire lock for {path}: {e}")
+            # Even if the lock is not acquired, try to create the directory
+            os.makedirs(path, exist_ok=True)
+
+        return path
 
     @staticmethod
     def get_rng_state():
