+++ verl/utils/distributed.py
@@ -23,7 +23,7 @@ from verl.utils.device import get_nccl_backend, get_torch_device
 def initialize_global_process_group(timeout_second=36000):
     from datetime import timedelta
 
-    torch.distributed.init_process_group(get_nccl_backend(), timeout=timedelta(seconds=timeout_second), init_method=os.environ.get("DIST_INIT_METHOD", None))
+    torch.distributed.init_process_group(get_nccl_backend(), timeout=timedelta(seconds=timeout_second))
     local_rank = int(os.environ["LOCAL_RANK"])
     rank = int(os.environ["RANK"])
     world_size = int(os.environ["WORLD_SIZE"])
