+++ verl/utils/flops_counter.py
@@ -16,7 +16,7 @@ from transformers import PretrainedConfig
 
 from verl.utils.device import get_torch_device
 
-VALID_CONFIG_TYPE = {"llama", "qwen2", "qwen2_vl", "qwen2_5_vl", "qwen3", "qwen3_moe", "deepseek_v3", "minicpmv", "minicpmo"}
+VALID_CONFIG_TYPE = {"llama", "qwen2", "qwen2_vl", "qwen2_5_vl", "qwen3", "qwen3_moe", "deepseek_v3"}
 
 
 def get_device_flops(unit="T"):
@@ -70,14 +70,11 @@ class FlopsCounter:
         self.estimate_func = {
             "qwen2": self._estimate_qwen2_flops,
             "llama": self._estimate_qwen2_flops,
-            "qwen2_moe": self._estimate_qwen2_moe_flops,
             "qwen2_vl": self._estimate_qwen2_flops,
             "qwen2_5_vl": self._estimate_qwen2_flops,
             "qwen3": self._estimate_qwen2_flops,
-            "qwen3_moe": self._estimate_qwen2_moe_flops,
+            "qwen3_moe": self._estimate_qwen3_moe_flops,
             "deepseek_v3": self._estimate_deepseek_v3_flops,
-            "minicpmv": self._estimate_qwen2_flops,
-            "minicpmo": self._estimate_qwen2_flops,
         }
         self.config = config
 
@@ -164,7 +161,7 @@ class FlopsCounter:
 
         return flops_achieved
 
-    def _estimate_qwen2_moe_flops(self, tokens_sum, batch_seqlens, delta_time):
+    def _estimate_qwen3_moe_flops(self, tokens_sum, batch_seqlens, delta_time):
         hidden_size = self.config.hidden_size
         vocab_size = self.config.vocab_size
         num_hidden_layers = self.config.num_hidden_layers
