+++ verl/utils/fs.py
@@ -252,34 +252,3 @@ def copy_local_path_from_hdfs(src: str, cache_dir=None, filelock=".file.lock", v
         return local_path
     else:
         return src
-
-
-def local_mkdir_safe(path):
-    """_summary_
-    Thread-safe directory creation function that ensures the directory is created
-    even if multiple processes attempt to create it simultaneously.
-
-    Args:
-        path (str): The path to create a directory at.
-    """
-
-    from filelock import FileLock
-
-    if not os.path.isabs(path):
-        working_dir = os.getcwd()
-        path = os.path.join(working_dir, path)
-
-    # Using hash value of path as lock file name to avoid long file name
-    lock_filename = f"ckpt_{hash(path) & 0xFFFFFFFF:08x}.lock"
-    lock_path = os.path.join(tempfile.gettempdir(), lock_filename)
-
-    try:
-        with FileLock(lock_path, timeout=60):  # Add timeout
-            # make a new dir
-            os.makedirs(path, exist_ok=True)
-    except Exception as e:
-        print(f"Warning: Failed to acquire lock for {path}: {e}")
-        # Even if the lock is not acquired, try to create the directory
-        os.makedirs(path, exist_ok=True)
-
-    return path
