+++ verl/workers/rollout/vllm_rollout/vllm_rollout_spmd.py
@@ -326,8 +326,6 @@ class vLLMRollout(BaseRollout):
                     non_tensor_batch["tools_kwargs"] = _repeat_interleave(non_tensor_batch["tools_kwargs"], self.sampling_params.n)
                 if "interaction_kwargs" in non_tensor_batch.keys():
                     non_tensor_batch["interaction_kwargs"] = _repeat_interleave(non_tensor_batch["interaction_kwargs"], self.sampling_params.n)
-                if "raw_prompt" in non_tensor_batch.keys():
-                    non_tensor_batch["raw_prompt"] = _repeat_interleave(non_tensor_batch["raw_prompt"], self.sampling_params.n)
 
             seq = torch.cat([idx, response], dim=-1)
